+++ R/plot.BchronologyRun.R
@@ -26,6 +26,7 @@
 #' @importFrom ggridges geom_ridgeline
 #' @importFrom scales pretty_breaks
 #' @importFrom stringr str_pad
+#' @importFrom grDevices rgb
 #'
 #' @export
 plot.BchronologyRun <-
@@ -33,7 +34,7 @@ plot.BchronologyRun <-
            dateHeight = 100,
            dateLabels = TRUE,
            dateLabelSize = 2,
-           dateCol = "darkslategray",
+           dateCol = rgb(47 / 255, 79 / 255, 79 / 255, 0.5),
            chronCol = "deepskyblue4",
            chronTransparency = 0.75,
            alpha = 0.95,