+++ tests/testthat/test_BchronCalibrate.R
@@ -1,6 +1,6 @@
-context("BchronCalibrate and related functions")
-
+set.seed(123)
 library(Bchron)
+co <- function(expr) capture.output(expr, file = "NUL")
 
 ages1 <- BchronCalibrate(
   ages = 11553,
@@ -40,10 +40,31 @@ test_that("BchronCalibrate works", {
   expect_s3_class(ages3, "BchronCalibratedDates")
 })
 
-test_that("Out of range dates produce a warning", {
+test_that("Out of range dates produce an error unless turned off", {
   expect_error(BchronCalibrate(
     ages = 85,
     ageSds = 30,
+    calCurve = "intcal20"
+  ))
+  expect_s3_class(
+    BchronCalibrate(
+      ages = 85,
+      ageSds = 30,
+      calCurve = "intcal20",
+      allowOutside = TRUE
+    ), "BchronCalibratedDates"
+  )
+})
+
+test_that("Ages converted to integers gives a message", {
+  expect_message(BchronCalibrate(
+    ages = 1500.3,
+    ageSds = 30,
+    calCurve = "intcal20"
+  ))
+  expect_message(BchronCalibrate(
+    ages = 1500,
+    ageSds = 30.3,
     calCurve = "intcal20"
   ))
 })
@@ -55,17 +76,24 @@ test_that("summary.BchronologyRun works", {
 })
 
 test_that("plot.BchronCalibrate works", {
-  expect_s3_class(plot(ages1), "ggplot")
+  p <- plot(ages1)
+  expect_s3_class(p, "ggplot")
   expect_type(plot(ages2), "list")
-  expect_s3_class(plot(ages2, date = 1), "ggplot")
-  expect_s3_class(plot(ages2, date = "Date1"), "ggplot")
+  p <- plot(ages2, date = 1)
+  expect_s3_class(p, "ggplot")
+  p <- plot(ages2, date = "Date1")
+  expect_s3_class(p, "ggplot")
   expect_type(plot(ages2, date = c("Date1", "Date2")), "list")
   expect_error(plot(ages2, date = "5"))
   expect_error(plot(ages2, date = 7))
-  expect_s3_class(plot(ages3), "ggplot")
-  expect_s3_class(plot(ages4), "ggplot")
-  expect_s3_class(plot(ages1, includeCal = TRUE), "ggplot")
+  p <- plot(ages3)
+  expect_s3_class(p, "ggplot")
+  p <- plot(ages4)
+  expect_s3_class(p, "ggplot")
+  p <- plot(ages1, includeCal = TRUE)
+  expect_s3_class(p, "ggplot")
   expect_error(plot(ages3, includeCal = TRUE))
-  expect_s3_class(plot(ages5, includeCal = TRUE), "ggplot")
+  p <- plot(ages5, includeCal = TRUE)
+  expect_s3_class(p, "ggplot")
   expect_error(plot(ages2, includeCal = TRUE))
 })