+++ tests/testthat/test_RSL.R
@@ -1,27 +1,30 @@
-context("RSL functions")
-
+set.seed(123)
 library(Bchron)
 library(ggplot2)
+co <- function(expr) capture.output(expr, file = "NUL")
+
 data(TestChronData)
 data(TestRSLData)
 
-RSLchron <- with(
-  TestChronData,
-  Bchronology(
-    ages = ages,
-    ageSds = ageSds,
-    positions = position,
-    positionThicknesses = thickness,
-    ids = id,
-    calCurves = calCurves,
-    jitterPositions = TRUE,
-    predictPositions = TestRSLData$Depth,
-    iterations = 100,
-    burn = 20,
-    thin = 1
+co(
+  RSLchron <- with(
+    TestChronData,
+    Bchronology(
+      ages = ages,
+      ageSds = ageSds,
+      positions = position,
+      positionThicknesses = thickness,
+      ids = id,
+      calCurves = calCurves,
+      jitterPositions = TRUE,
+      predictPositions = TestRSLData$Depth,
+      iterations = 100,
+      burn = 20,
+      thin = 1
+    )
   )
 )
-RSLrun <- with(
+co(RSLrun <- with(
   TestRSLData,
   BchronRSL(RSLchron,
     RSLmean = RSL,
@@ -31,7 +34,7 @@ RSLrun <- with(
     burn = 20,
     thin = 1
   )
-)
+))
 
 test_that("RSL Bchronology", {
   expect_s3_class(RSLchron, "BchronologyRun")
@@ -42,14 +45,14 @@ test_that("RSL functions", {
 })
 
 test_that("summary and plot RSL functions", {
-  expect_s3_class(
-    plot(RSLrun, type = "RSL") + ggtitle("Relative sea level plot"),
-    "ggplot"
-  )
-  expect_s3_class(plot(RSLrun, type = "rate") + ggtitle("Rate of RSL change") +
-    ylab("Rate (mm per year)"), "ggplot")
-  expect_s3_class(plot(RSLrun, type = "accel") + ggtitle("Rate of RSL change") +
-    ylab("Rate (mm per year)"), "ggplot")
+  p <- plot(RSLrun, type = "RSL") + ggtitle("Relative sea level plot")
+  expect_s3_class(p, "ggplot")
+  p <- plot(RSLrun, type = "rate") + ggtitle("Rate of RSL change") +
+    ylab("Rate (mm per year)")
+  expect_s3_class(p, "ggplot")
+  p <- plot(RSLrun, type = "accel") + ggtitle("Rate of RSL change") +
+    ylab("Rate (mm per year)")
+  expect_s3_class(p, "ggplot")
   expect_output(summary(RSLrun, type = "RSL", age_grid = seq(0, 2000, by = 250)))
   expect_output(summary(RSLrun, type = "parameters", age_grid = seq(0, 2000, by = 250)))
   expect_output(summary(RSLrun, type = "rate", age_grid = seq(0, 2000, by = 250)))